drafts/sockets/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.3
chardet==4.0.0
click==8.0.3
dnspython==2.1.0
eventlet==0.32.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
python-engineio==4.2.1
python-socketio==5.4.1
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.2
yarl==1.7.0
